Manually bind ParamFlags
Remove meaningless flags and add user-defined flags.
