+++ glib/Gir_GObject.toml
@@ -12,7 +12,6 @@ trust_return_value_nullability = true
 
 generate = [
     "GObject.BindingFlags",
-    "GObject.ParamFlags",
     "GObject.SignalFlags",
 ]
 
@@ -22,6 +21,7 @@ ignore = [
 manual = [
     "GObject.Object",
     "GObject.Value",
+    "GObject.ParamFlags",
 ]
 
 [[object]]
