+++ glib/src/gobject/auto/flags.rs
@@ -69,45 +69,6 @@ impl SetValue for BindingFlags {
 }
 
 bitflags! {
-    pub struct ParamFlags: u32 {
-        const READABLE = 1;
-        const WRITABLE = 2;
-        const READWRITE = 3;
-        const CONSTRUCT = 4;
-        const CONSTRUCT_ONLY = 8;
-        const LAX_VALIDATION = 16;
-        const STATIC_NAME = 32;
-        const PRIVATE = 32;
-        const STATIC_NICK = 64;
-        const STATIC_BLURB = 128;
-        const EXPLICIT_NOTIFY = 1073741824;
-        const DEPRECATED = 2147483648;
-    }
-}
-
-impl fmt::Display for ParamFlags {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        <Self as fmt::Debug>::fmt(self, f)
-    }
-}
-
-#[doc(hidden)]
-impl ToGlib for ParamFlags {
-    type GlibType = gobject_ffi::GParamFlags;
-
-    fn to_glib(&self) -> gobject_ffi::GParamFlags {
-        self.bits()
-    }
-}
-
-#[doc(hidden)]
-impl FromGlib<gobject_ffi::GParamFlags> for ParamFlags {
-    unsafe fn from_glib(value: gobject_ffi::GParamFlags) -> ParamFlags {
-        ParamFlags::from_bits_truncate(value)
-    }
-}
-
-bitflags! {
     pub struct SignalFlags: u32 {
         const RUN_FIRST = 1;
         const RUN_LAST = 2;
