+++ glib/src/gobject/auto/mod.rs
@@ -7,7 +7,6 @@ pub use self::binding::Binding;
 
 mod flags;
 pub use self::flags::BindingFlags;
-pub use self::flags::ParamFlags;
 pub use self::flags::SignalFlags;
 
 #[doc(hidden)]
